Move groovy stuff to its own feature variant
