+++ src/groovy/java/dev/lukebemish/codecextras/groovy/structured/reflective/implementation/package-info.java
@@ -1,0 +1,6 @@
+@NullMarked
+@ApiStatus.Internal
+package dev.lukebemish.codecextras.groovy.structured.reflective.implementation;
+
+import org.jetbrains.annotations.ApiStatus;
+import org.jspecify.annotations.NullMarked;
