+++ src/main/java/module-info.java
@@ -1,3 +1,5 @@
+import dev.lukebemish.codecextras.structured.reflective.implementation.BuiltInReflectiveStructureCreator;
+
 module dev.lukebemish.codecextras {
     uses dev.lukebemish.codecextras.companion.AlternateCompanionRetriever;
     uses dev.lukebemish.codecextras.structured.reflective.ReflectiveStructureCreator;
@@ -15,7 +17,6 @@ module dev.lukebemish.codecextras {
     requires static org.objectweb.asm;
     requires static org.slf4j;
     requires static com.google.auto.service;
-    requires static org.apache.groovy;
 
     exports dev.lukebemish.codecextras;
     exports dev.lukebemish.codecextras.comments;
@@ -39,4 +40,6 @@ module dev.lukebemish.codecextras {
     exports dev.lukebemish.codecextras.types;
 
     exports dev.lukebemish.codecextras.internal to codecextras_minecraft, dev.lukebemish.codecextras.minecraft;
+
+    provides dev.lukebemish.codecextras.structured.reflective.ReflectiveStructureCreator with BuiltInReflectiveStructureCreator;
 }
